+++ examples/WireMock.Net.Console.Proxy.Net452/Program.cs
@@ -1,4 +1,6 @@
-﻿using Newtonsoft.Json;
+﻿using System;
+using System.Net.Http;
+using Newtonsoft.Json;
 using WireMock.Server;
 using WireMock.Settings;
 
@@ -8,14 +10,15 @@ namespace WireMock.Net.Console.Proxy.Net452
     {
         static void Main(string[] args)
         {
+            string[] urls = { "http://localhost:9091/", "https://localhost:9443/" };
             var server = FluentMockServer.Start(new FluentMockServerSettings
             {
-                Urls = new[] { "http://localhost:9091/", "https://localhost:9443/" },
+                Urls = urls,
                 StartAdminInterface = true,
                 ReadStaticMappings = false,
                 ProxyAndRecordSettings = new ProxyAndRecordSettings
                 {
-                    Url = "https://www.google.com",
+                    Url = "http://postman-echo.com/post",
                     //ClientX509Certificate2ThumbprintOrSubjectName = "www.yourclientcertname.com OR yourcertificatethumbprint (only if the service you're proxying to requires it)",
                     SaveMapping = true,
                     SaveMappingToFile = false,
@@ -27,6 +30,13 @@ namespace WireMock.Net.Console.Proxy.Net452
             {
                 System.Console.WriteLine(JsonConvert.SerializeObject(eventRecordArgs.NewItems, Formatting.Indented));
             };
+
+            var uri = new Uri(urls[0]);
+            var form = new MultipartFormDataContent
+            {
+                { new StringContent("data"), "test", "test.txt" }
+            };
+            new HttpClient().PostAsync(uri, form).GetAwaiter().GetResult();
 
             System.Console.WriteLine("Press any key to stop the server");
             System.Console.ReadKey();