+++ src/WireMock.Net/ResponseMessageBuilder.cs
@@ -9,7 +9,10 @@ namespace WireMock
     internal static class ResponseMessageBuilder
     {
         private static string ContentTypeJson = "application/json";
-        private static readonly IDictionary<string, WireMockList<string>> ContentTypeJsonHeaders = new Dictionary<string, WireMockList<string>> { { HttpKnownHeaderNames.ContentType, new WireMockList<string> { ContentTypeJson } } };
+        private static readonly IDictionary<string, WireMockList<string>> ContentTypeJsonHeaders = new Dictionary<string, WireMockList<string>>
+        {
+            { HttpKnownHeaderNames.ContentType, new WireMockList<string> { ContentTypeJson } }
+        };
 
         internal static ResponseMessage Create(string message, int statusCode = 200, Guid? guid = null)
         {